Chore: change path & layout
