+++ src/app/(NoLayout)/layout.tsx
@@ -1,0 +1,19 @@
+import type { Metadata } from 'next';
+import '../globals.css';
+
+export const metadata: Metadata = {
+  title: 'KT',
+  description: '',
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="ko">
+      <body>{children}</body>
+    </html>
+  );
+}
